+++ RSACipher/RSACipher/RSAkey.cs
@@ -1,13 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Numerics;
 
 namespace RSACipher
 {
     public class RSAkey
     {
-        public int modulus;
-        public int power;
+        public BigInteger modulus;
+        public BigInteger power;
 
     }
 }
